# src/CollectorServer/src/IO.Swagger/Models/EmailSchema.cs
/*
 * Communication Collector
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class EmailSchema : IdentityEmail, IEquatable<EmailSchema>
    { 
        /// <summary>
        /// Gets or Sets Newsletter
        /// </summary>

        [DataMember(Name="newsletter")]
        public bool? Newsletter { get; set; }

        /// <summary>
        /// Gets or Sets Deals
        /// </summary>

        [DataMember(Name="deals")]
        public bool? Deals { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class EmailSchema {\n");
            sb.Append("  Newsletter: ").Append(Newsletter).Append("\n");
            sb.Append("  Deals: ").Append(Deals).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public  new string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((EmailSchema)obj);
        }

        /// <summary>
        /// Returns true if EmailSchema instances are equal
        /// </summary>
        /// <param name="other">Instance of EmailSchema to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(EmailSchema other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Newsletter == other.Newsletter ||
                    Newsletter != null &&
                    Newsletter.Equals(other.Newsletter)
                ) && 
                (
                    Deals == other.Deals ||
                    Deals != null &&
                    Deals.Equals(other.Deals)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Newsletter != null)
                    hashCode = hashCode * 59 + Newsletter.GetHashCode();
                    if (Deals != null)
                    hashCode = hashCode * 59 + Deals.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(EmailSchema left, EmailSchema right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(EmailSchema left, EmailSchema right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}